metehan/api/requirements.txt
fastapi
uvicorn
mangum
psycopg2-binary
python-dotenv
requests
pydantic
boto3
python-multipart==0.0.9
ultralytics